September 11, 2023 - Energy Tabler is Done
Major push containing the following changes:

Added an unzipped version of Unitrans_GTFS_raw.  The folder contains the individual txt files of the component parts of the GTFS feed.  Copying the file and changing the extension to .csv will transform them into CSV files if the need to be worked with in R or other programs.

Added a version of Unitrans_GTFS that has fixed a small error in the feed.  This error is inconsequential to the model, but it does cause warnings to pop up when the feed is imported.

Code Changes:
R/utils.R

-Added jan_first as a function.  This is a helper function that generates a lubridate object of January 1st of the system year. This date is used when building the service calendar throughout the operating year.

R/energyTabler.R

-Fixed issue with year_tabler.  The function now works as intended.
-Added day_tabler.  This function is working as intended, and fills the year_tabler table with information from the data.frame generated by gtfs_framer().
--Should year_tabler() and day_tabler() be combined into a single function?
-Added gtfs_summarize().  The function is working as intended.  It generates the summary table that is used by energy_tabler() and saves it to the local directory for future reference.
-Added energy_tabler(). The function throws some warnings, but correctly generates an energy table for GTFS feeds as expected.