Major Content Commit - We're Shiny!
Modified energyTabler() for future use in the shiny app.  Now it can be called to return 0 (as before) or return the generated table.  Saving is now also optional.

Added shinyApp, which is still deep in development.

Added shinyUtils, which contains utility functions used exclusively in ZEBRa-Lite.

Added some toy data for use in testing certain parts of ZEBRa-Lite.

Small update to .gitignore.