Another significant update for Shiny
Added candLoc.R (probably to be renamed in the future) which contains the mapping functions for the cand_loc_searcher function as well as the function itself.

Added some utils to utils.R for cand_loc_searcher.

Some updates to shinyApp.R, but I can't get the map and candLoc table to render.