Another small commit
Fixed conflicts created by improper eariler commit in shinyApp.R

Small change to addNewRow in shinyUtils.R to have objects of an unknown class be added as `NULL` rather than `NA`.  This should not change overall functionality.

Slight reorganization in the front-end: if mapping functionality can be restored, this will be displayed in a 'results' tab, with the 'dashboard' tab reserved for quick-changing input values and the 'recalculate' button.
Note: I have gone back and forth on this design a number of times, so any suggestion on which is better is welcome.

Fixed issue when editing DTs in the shiny app.  The issue was caused by improper handling of the _cell_edit information when a full row was targeted.

Removed some testing text in the shiny app.