Significant push to setup future utility
This commit features a reorganization of the candLoc.R file in preparation for having multiple ways of choosing the "best" locations in the network for opportunity chargers.

Created mappers.R, which currently holds the gtfs_mapper() function (which currently does not work). Any future leaflet- or mapping-related functions should go here.

Added the function cand_loc_mode(), in preparation for use in the shinyApp.

Added two modes of cand_loc_searcher:
1. Repeats the default behavior of the cand_loc_searcher default app by ensuring that all routes have at least one opportunity charger.
2. Repeats the behavior of the cand_loc_searcher default app when all.routes was set to FALSE; the chargers are allocated based on where the most opportunities for charging exist (that is, which stops are stopped at the most).

Added functionality to the shinyApp to "grey out" the Opportunity Chargers input if mode 1 is selected, to better indicate to the user that this number is not used in mode 1.