First Major Commit in a While!
MAJOR CHANGES:
Renamed `get_routes_sldf` to `prep_routes_for_map`. This is a more descriptive name for the function, especially since `sp` has been sunsetted.  This function is still pretty 'dirty' and has a lot of code taken from the `gtfsr` package (which is also depricated at this point), so it needs to be modernized.  We can probably also pass the data back out of this function in the form of a single data.frame, rather than the environment that is currently returned.

Removing dependencies on `sp` has also fixed the mapping issues that cropped up earlier.  The mapping part of ZEBRa Lite works again.  Dependency was removed by changing the nature of the `gtfs_lines` data.frame.  It now uses the sf object 'LINESTRING' to store lat/lon information.  Leaflet is happy to use this kind of data in its mapping, and the overall implementation is cleaner than embedding code into a data.frame to be passed back out to the mapper function.

MINOR CHANGES:

The test output text has been moved to the Results tab for now, where it was used to check that the 'Calculate' button was being detected correctly.

Consolidated the 'Infrastructure Parameters' tab with the 'Dashboard' tab.  I might break this back out again if further costs are needed.

Added space for the user to input their Google Maps API key when using the API to generate an energy table.  Note that this functionality is not currently implemented.

The 'Calculate' button generates a map once again (see above).

TO DO:
Decide on some final outputs and begin writing calculation functions to reach them.

Clean up `prep_routes_for_map`.

Generate README and any additional documentation that may be required (likely using RMD).

Implement the modality of `cand_loc_searcher` and a way to use subsets of the network.  Could this be a checklist of routes?